requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pypng==0.20220715.0
Pillow==10.1.0
ipython==8.17.2
opencv-python==4.8.1.78